Fixes #1552 Update to .NET8
