+++ src/MoBi.Assets/MoBi.Assets.csproj
@@ -22,9 +22,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.33" /> 
-    <PackageReference Include="OSPSuite.Assets.Images" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.33" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.34" /> 
+    <PackageReference Include="OSPSuite.Assets.Images" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.34" />
   </ItemGroup>
 
   <ItemGroup>
