+++ src/MoBi.CLI/MoBi.CLI.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
 
-    <TargetFrameworks>net8;net472</TargetFrameworks>
+    <TargetFrameworks>net8</TargetFrameworks>
      <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
      <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
      <Authors>Open-Systems-Pharmacology</Authors>
@@ -22,17 +22,16 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.6" />
     <PackageReference Include="CommandLineParser" Version="2.9.1" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.2" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Presentation" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Assets" Version="13.0.33" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Presentation" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Assets" Version="13.0.34" />
     <PackageReference Include="OSPSuite.Utility" Version="4.1.1.1" />
-    <PackageReference Include="OSPSuite.FuncParser" Version="4.0.0.73" GeneratePathProperty="true" />
-    <PackageReference Include="OSPSuite.SimModel" Version="4.0.0.75" GeneratePathProperty="true" />
+    <PackageReference Include="OSPSuite.FuncParser" Version="4.0.0.74" GeneratePathProperty="true" />
+    <PackageReference Include="OSPSuite.SimModel" Version="4.0.0.76" GeneratePathProperty="true" />
     <PackageReference Include="OSPSuite.SimModelSolver_CVODES" Version="4.1.0.19" GeneratePathProperty="true" />
     <PackageReference Include="System.Data.SQLite.Core" Version="1.0.119" GeneratePathProperty="true" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.33" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.34" />
   </ItemGroup>
   <ItemGroup>
      <Content Include="..\..\dimensions\OSPSuite.Dimensions.xml" Link="OSPSuite.Dimensions.xml">
