+++ src/MoBi.Presentation/UICommand/ICopyTextCommand.cs
@@ -1,0 +1,9 @@
+using OSPSuite.Presentation.UICommands;
+
+namespace MoBi.Presentation.UICommand
+{
+   public interface ICopyTextCommand : IObjectUICommand<string>
+   {
+
+   }
+}
