+++ src/MoBi.Presentation/Views/BaseDiagram/IMoBiBaseDiagramView.cs
@@ -1,4 +1,3 @@
-using System.Windows.Forms;
 using MoBi.Presentation.Presenter.BaseDiagram;
 using OSPSuite.Core.Diagram;
 using OSPSuite.Presentation.Views;
@@ -10,6 +9,5 @@ namespace MoBi.Presentation.Views.BaseDiagram
    {
       void ExpandParents(IBaseNode baseNode);
       bool IsMoleculeNode(IBaseNode baseNode);
-      Control Overview { set; }
    }
 }
