+++ src/MoBi.UI/MoBi.UI.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
@@ -19,15 +19,15 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.2" />
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.2" />
-    <PackageReference Include="OSPSuite.DataBinding" Version="3.0.1.1" />
-    <PackageReference Include="OSPSuite.DataBinding.DevExpress" Version="6.0.1.2" />
-    <PackageReference Include="OSPSuite.DevExpress" Version="21.2.15" />
+    <PackageReference Include="OSPSuite.DataBinding" Version="3.0.1.2" />
+    <PackageReference Include="OSPSuite.DataBinding.DevExpress" Version="6.0.1.3" />
+    <PackageReference Include="OSPSuite.DevExpress" Version="21.2.15.1" />
     <PackageReference Include="OSPSuite.Utility" Version="4.1.1.1" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Presentation" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.UI" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.33" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Presentation" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.UI" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.34" />
    </ItemGroup>
   <ItemGroup>
     <None Include="..\..\LICENSE">
@@ -40,11 +40,6 @@
     <ProjectReference Include="..\MoBi.Assets\MoBi.Assets.csproj" />
     <ProjectReference Include="..\MoBi.Core\MoBi.Core.csproj" />
     <ProjectReference Include="..\MoBi.Presentation\MoBi.Presentation.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.VisualBasic" />
-    <Reference Include="System.Windows.Forms" />
   </ItemGroup>
 
   <ItemGroup>
