+++ src/MoBi.UI/UICommands/CopyTextCommand.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Forms;
+using MoBi.Presentation.UICommand;
+using OSPSuite.Presentation.UICommands;
+
+namespace MoBi.UI.UICommands
+{
+   public class CopyTextCommand : ObjectUICommand<string>, ICopyTextCommand
+   {
+      protected override void PerformExecute()
+      {
+         Clipboard.SetText(Subject);
+      }
+   }
+}
