+++ tests/MoBi.HelpersForTests/MoBi.HelpersForTests.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="OSPSuite.Core" Version="13.0.33" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.34" />
   </ItemGroup>
 
   <ItemGroup>
