+++ tests/MoBi.Tests/MoBi.Tests.csproj
@@ -1,18 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-
+    <TargetFramework>net8.0-windows</TargetFramework>
     <IsPackable>false</IsPackable>
-
     <RootNamespace>MoBi</RootNamespace>
     <Version Condition="'$(Version)' == ''">13.0.0</Version>
   </PropertyGroup>
-
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
     <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
-
   <ItemGroup>
     <Content Include="..\..\dimensions\OSPSuite.Dimensions.xml" Link="OSPSuite.Dimensions.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
@@ -33,7 +28,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="FakeItEasy" Version="8.3.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="5.0.0">
@@ -43,17 +37,15 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
     <PackageReference Include="nunit" Version="3.14.0" />
     <PackageReference Include="OSPSuite.BDDHelper" Version="4.0.1.1" />
-    <PackageReference Include="OSPSuite.Core" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.UI" Version="13.0.33" />
-    <PackageReference Include="OSPSuite.FuncParser" Version="4.0.0.73" GeneratePathProperty="true" />
-    <PackageReference Include="OSPSuite.SimModel" Version="4.0.0.75" GeneratePathProperty="true" />
+    <PackageReference Include="OSPSuite.Core" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Infrastructure" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.Infrastructure.Castle" Version="13.0.34" />
+    <PackageReference Include="OSPSuite.FuncParser" Version="4.0.0.74" GeneratePathProperty="true" />
+    <PackageReference Include="OSPSuite.SimModel" Version="4.0.0.76" GeneratePathProperty="true" />
     <PackageReference Include="OSPSuite.SimModelSolver_CVODES" Version="4.1.0.19" GeneratePathProperty="true" />
     <PackageReference Include="System.Data.SQLite.Core" Version="1.0.119" GeneratePathProperty="true" />
     <PackageReference Include="OSPSuite.TeXReporting" Version="3.0.1.1" GeneratePathProperty="true" />
   </ItemGroup>
-
   <ItemGroup>
     <ProjectReference Include="..\..\src\MoBi.CLI.Core\MoBi.CLI.Core.csproj" />
     <ProjectReference Include="..\..\src\MoBi.Core\MoBi.Core.csproj" />
@@ -64,9 +56,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Reference Include="libsbmlcsP">
-      <HintPath>..\..\pre\libsbmlcsP.dll</HintPath>
-    </Reference>
+     <Reference Include="libsbmlcsP">
+        <HintPath>..\..\pre\libsbmlcsP.dll</HintPath>
+     </Reference>
   </ItemGroup>
 
 
@@ -88,13 +80,9 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-
-
   <ItemGroup>
     <Folder Include="ProjectConversion\v6_0\" />
   </ItemGroup>
-
-
   <ItemGroup>
     <None Update="Core\SBML\Testfiles\Barros2021_RAJI.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
@@ -106,7 +94,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-
   <!-- Content comes from Stub.system.data.sqlite.core. Basically, implements the same task as if you added this binary as content and copy to output folder -->
   <PropertyGroup>
     <ContentSQLiteInteropFiles>true</ContentSQLiteInteropFiles>
@@ -114,11 +101,4 @@
     <CleanSQLiteInteropFiles>false</CleanSQLiteInteropFiles>
     <CollectSQLiteInteropFiles>false</CollectSQLiteInteropFiles>
   </PropertyGroup>
-
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <ItemGroup>
-      <WindowsFiles Include="$(TargetDir)x64/SQLite.Interop.dll" />
-    </ItemGroup>
-    <Copy SourceFiles="@(WindowsFiles);" DestinationFolder="$(TargetFolder)" DestinationFiles="@(WindowsFiles-&gt;Replace('x64/SQLite.Interop.dll', 'SQLite.Interop.dll'))" />
-  </Target>
 </Project>
