+++ tests/MoBi.Tests/Presentation/Mapper/InitialConditionToObjectBaseSummaryDTOMapperSpecs.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using System.Runtime.Remoting.Contexts;
 using MoBi.Assets;
 using OSPSuite.BDDHelper;
 using OSPSuite.BDDHelper.Extensions;
